Login with Google Added

+++ server/routers/memoryRouter.js
@@ -48,7 +48,8 @@ router.put("/:id", auth, async (req, res) => {
 
     if(!mongoose.Types.ObjectId.isValid(id))
       res.status(404).json({message: 'Memory id is not valid'})
-  
+    const oldMemory = await Memory.findById(id)
+    if(req.creatorId !== oldMemory.creatorId) return res.status(403)
     const {title, content, creator,image} = req.body
     
     const updatedMemory = await Memory.findByIdAndUpdate(id,{title, content, creator,image, _id: id},{new:true})
@@ -64,6 +65,8 @@ router.delete("/:id", auth, async (req, res) => {
 
     if(!mongoose.Types.ObjectId.isValid(id))
       res.status(404).json({message: 'Memory id is not valid'})
+    const oldMemory = await Memory.findById(id)
+    if(req.creatorId !== oldMemory.creatorId) return res.status(403)
 
     await Memory.findByIdAndDelete(id)
     res.status(200).json({message: 'Memory has been deleted'})
